Rename the ui directory to view.
