+++ src/cmapps/neon/neon.py
@@ -18,7 +18,7 @@ import sys
 from PySide6 import QtWidgets
 
 from cmapps.neon.core.mainapplication import MainApplication
-from cmapps.neon.ui.mainwindow import MainWindow
+from cmapps.neon.view.mainwindow import MainWindow
 from cmapps.neon.settings.mainsettings import set_application_settings
