+++ src/cmapps/neon/view/icons_rc.py
@@ -701,7 +701,7 @@ l$\x82|^k\xba\x17\xf9{\x016\x12\x99$6\
 \xbd*\xd8\xd5\x89m#v?<\x7f\xf3\xa6UhH\
 \xdc\x8f\xc2|\x0dq\xdd\x8fH\xc7:l\xdd\xd6\x8cK\
 \x1e\xce\xc3\xf2\x8d\xde^I\xbel\xf7\x15=\x8a\x07.\
-\xe8\xc4.#ui\x9c\x9eh\xdeM\x81\xdf\xd2\xe5\xc7\
+\xe8\xc4.#view\x9c\x9eh\xdeM\x81\xdf\xd2\xe5\xc7\
 \x8d\xcf\xec\x8d\x96\xee\x90\xe5\xd5\xbf\xfc\xd2\xcb\x19\xc1\xb7\
 \xa6`UTk\xe6\x11\xea\xcd\x93\x14\xcd\xb4l\xff\x91\
 G\x1fIs\xf8\xdc|\x01'\xb0n\xd3\xd0\xc8 ;\
