+++ src/cmapps/neon/view/misc/scientificdoublespinbox.py
